requirements.txt
google-generativeai==0.7.2
deepdiff==7.0.1
python-dotenv==1.0.1
requests==2.32.3
faster-whisper==1.0.3
